requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
groq>=0.11.0
python-dotenv==1.0.1
pytest==8.3.3
nltk>=3.8.1